# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Little Miracles API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://littlemiracles.test/api",
            "description": "Development API"
        },
        {
            "url": "https://littlemiracles.viitech.net/api",
            "description": "Staging API"
        },
        {
            "url": "https://api.littlemiraclesbys.com/api",
            "description": "Production API"
        }
    ],
    "paths": {
        "/api/data": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "App Data",
                "description": "App Data",
                "operationId": "2f5eb0163e3f87d089135347b76a04ba",
                "parameters": [
                    {
                        "name": "last_update",
                        "in": "query",
                        "description": "Last Update: 2020-10-04",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomJsonResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomJsonResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/available-hours": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "List All Available Hours",
                "description": "List of Available Hours",
                "operationId": "2a4ff53b9427a4ba5a33f1ccf5aab1bf",
                "parameters": [
                    {
                        "name": "last_update",
                        "in": "query",
                        "description": "Last Update: 2020-10-04",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date: 2020-10-04",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available hours retrived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomJsonResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomJsonResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CustomJsonResponse": {
                "description": "Class CustomJsonResponse",
                "properties": {
                    "message": {
                        "description": "Message",
                        "type": "string"
                    },
                    "data": {
                        "description": "Data",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "error": {
                        "description": "Error",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}